SMILES-Demo/requirements.txt
joblib==1.1.0
matplotlib==3.4.3
networkx==2.6.3
numpy==1.20.3
pandas==1.3.4
Pillow==9.0.1
plotly==5.6.0
py3Dmol==1.8.0
pysmiles==1.0.1
rdkit_pypi==2021.9.1
requests==2.26.0
seaborn==0.11.2
setuptools==58.0.4
stmol==0.0.7
